Differential Revision: D15793003
Differential Version: 85071961

+++ aten/src/ATen/native/quantized/cpu/qlinear_prepack.cpp
@@ -75,7 +75,7 @@ class QLinearPackWeightInt8 final : public c10::OperatorKernel {
             /*pmat=*/nullptr, // PackBMatrix manages ownership of pmat
             /*groups=*/1),
         col_offsets,
-        static_cast<float>(weight.q_scale()),
+        weight.q_scale(),
         weight_zero_point_int32});
 
     // TODO: we will need to replace this with torchscript classes at a later
